requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
blinker==1.8.2
bottle==0.13.1
cachelib==0.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
clr-loader==0.2.6
colorama==0.4.6
dash==2.18.1
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-extensions==1.0.18
dash-html-components==2.0.0
dash-table==5.0.0
dataclass-wizard==0.22.3
EditorConfig==0.12.4
Flask==3.0.3
Flask-Caching==2.3.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsbeautifier==1.15.1
MarkupSafe==2.1.5
more-itertools==10.5.0
nest-asyncio==1.6.0
numpy==2.1.1
orjson==3.10.7
packaging==24.1
pandas==2.2.3
pefile==2024.8.26
pip-licenses==5.0.0
plotly==5.24.1
plotly-resampler==0.10.0
prettytable==3.11.0
proxy-tools==0.1.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
pyTDFSDK @ git+https://github.com/gtluu/pytdfsdk@4808da81d3e37f82da96f392dbb446c3733edacb
python-dateutil==2.9.0.post0
pythonnet==3.0.4
pytz==2024.2
pywebview==5.2
pywin32-ctypes==0.2.3
requests==2.32.3
retrying==1.3.4
ruff==0.4.10
six==1.16.0
tenacity==9.0.0
tomli==2.0.1
tsdownsample==0.1.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.0.4
zipp==3.20.2
